requirements/base.txt
Django==1.6.8
Markdown==2.5.2
Unipath==1.0
django-filter==0.9.1
djangorestframework==3.0.0
